+++ Sources/Capsomnia/main.swift
@@ -1,6 +1,7 @@
 import AppKit
 
 let app = NSApplication.shared
+app.appearance = NSAppearance(named: .darkAqua)
 let delegate = Capsomnia()
 app.delegate = delegate
 app.run()